Generation Component Automatically Fetches Data
The Generation component now automatically fetches the next generation recursively, using the calculated time from the current generation's expiration as the delay. 

The delay is set to a default time if the expiration is too short, in order to restrain the number of recursive calls and help prevent the application from freezing.

The Timeout is cleared when the component will be unmounted, to prevent it from redundantly running in the background.